* Moved ReportsCommand.java * Fix regex from misfiring
